Use single resource
