+++ libraries/provider_chatops_notify_custom.rb
@@ -3,7 +3,8 @@ require_relative 'helpers'
 
 class Chef
   class Provider
-    class CustomNotify < Chef::Provider::LWRPBase
+    class ChatopsNotifyCustom < Chef::Provider::LWRPBase
+      provides :chatops_notify
       include Helpers::Http
       use_inline_resources if defined?(use_inline_resources)
 
@@ -16,6 +17,13 @@ class Chef
           body = new_resource.body.to_json
           http_uri(body)
           req.body = new_resource.body.to_json
+        end
+      end
+
+      class << self
+        # supports the given resource and action (late binding)
+        def supports?(resource, _action)
+          resource.chat_platform == :custom
         end
       end
     end
